pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codeguard-ai"
version = "1.0.0"
description = "AI-powered Codebase Vulnerability Scanner, Reviewer, and Auto-Commenter using RAG"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "CodeGuard AI Team"}
]
keywords = ["security", "vulnerability", "code-review", "rag", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlmodel>=0.0.14",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.12.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "pygithub>=1.59.0",
    "gitpython>=3.1.40",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "sentence-transformers>=2.3.0",
    "qdrant-client>=1.7.0",
    "tiktoken>=0.5.1",
    "openai>=1.3.0",
    "ollama>=0.1.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.2",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "rich>=13.7.0",
    "click>=8.1.7",
    "aiofiles>=23.2.1",
    "reportlab>=4.0.7",
    "weasyprint>=60.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "bandit>=1.7.5",
    "semgrep>=1.45.0",
    "safety>=2.3.5",
    "playwright>=1.40.0",
    "black>=23.11.0",
    "isort>=5.12.0",
]

[project.scripts]
codeguard = "app.cli:main"

[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "--cov=app --cov-report=html --cov-report=term-missing --cov-fail-under=90"

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

